Bug 1153460 - Followup to fix warnings-as-errors build failure on a CLOSED TREE. r=bustage.
UltraBlame original commit: d19d1522ce46c7ffaceabe17a047d425554ab324

+++ gfx/thebes/gfxCoreTextShaper.cpp
@@ -170,7 +170,7 @@ gfxCoreTextShaper::ShapeText(gfxContext      *aContext,
             (CTRunRef)::CFArrayGetValueAtIndex(glyphRuns, runIndex);
         
         CFRange range = ::CTRunGetStringRange(aCTRun);
-        if (range.location + range.length <= startOffset ||
+        if (uint32_t(range.location + range.length) <= startOffset ||
             range.location - startOffset >= aLength) {
             continue;
         }
